+++ CHANGELOG.md
@@ -18,6 +18,16 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Breaking
 
+## 2023-10-19
+
+- yellowstone-grpc-tools-1.0.0-rc.5+solana.1.17.1
+
+### Features
+
+- tools: add Google Pub/Sub ([#211](https://github.com/rpcpool/yellowstone-grpc/pull/211)).
+
+### Breaking
+
 - kafka: rename to tools ([#203](https://github.com/rpcpool/yellowstone-grpc/pull/203)).
 
 ## 2023-10-12